03_audio_websocket_pipecat/requirements.txt
pipecat-ai[silero,openai,deepgram,elevenlabs,websocket]
fastapi
uvicorn[standard]
python-dotenv
anthropic
